// src/types.ts
import {
	type CallToolResult,
	type Tool,
	ToolSchema,
} from '@modelcontextprotocol/sdk/types.js';
import type { z } from 'zod';

const ToolInputSchema = ToolSchema.shape.inputSchema;
export type ToolInput = z.infer<typeof ToolInputSchema>;

export type ToolRegistration<T> = Tool & {
	handler: (args: T) => CallToolResult | Promise<CallToolResult>;
};

export type InspectorApp = {
	id: string;
	title: string;
	description: string;
	type: string;
	devtoolsFrontendUrl: string;
	webSocketDebuggerUrl: string;
	deviceName: string;
	reactNative: {
		logicalDeviceId: string;
		capabilities: {
			prefersFuseboxFrontend: boolean;
			nativeSourceCodeFetching: boolean;
			nativePageReloads: boolean;
		};
	};
};